Create Fibonacci Operate
The script will display an animated bar chart showing the transformation of the Fibonacci numbers and count of each result, for both the original and reversed rules. It will update as new Fibonacci numbers are calculated.


each Fibonacci number is written in binary then under these conditions, a condition is created to break each digit to a string where
 1=1 and 0= [+] or [-]
If the number of 0's bisecting an even number of ones is even, then insert [-] where the zeros are. 

For example,
1001=> [ 1,0,0,1 ] => 1 [0,0] 1 => 1 [-] 1 

However, if the number of zeros between two 1's is odd then the space between 2 1's is occupied by a plus sign. 

For example,
101=> [1,0,1] => [1 [0] 1] => 1+1

All trailing zeros are treated as NaN values and are omitted from the code.

Then finally, the output is converted back to decimal and plotted on a number line in  bar chart